server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
torch>=2.0.0
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

